add ls persist to all forms
